[codex] Port v2 terminal hotkeys to v1 (#3724)
* port v2 terminal hotkeys to v1

* Fix terminal hotkey PR feedback
