+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/Terminal/helpers.ts
@@ -8,12 +8,6 @@ import { Unicode11Addon } from "@xterm/addon-unicode11";
 import { WebglAddon } from "@xterm/addon-webgl";
 import type { ITheme } from "@xterm/xterm";
 import { Terminal as XTerm } from "@xterm/xterm";
-import {
-	getBinding,
-	isTerminalReservedEvent,
-	matchesChord,
-	resolveHotkeyFromEvent,
-} from "renderer/hotkeys";
 import type { DetectedLink } from "renderer/lib/terminal/links";
 import { TerminalLinkManager } from "renderer/lib/terminal/terminal-link-manager";
 import { installRectangleRendererAlphaPatch } from "renderer/lib/terminal/webgl-vibrancy-patch";
@@ -24,10 +18,6 @@ import {
 	DEFAULT_THEME_ID,
 	getTerminalColors,
 } from "shared/themes";
-import {
-	shouldBubbleClipboardShortcut,
-	shouldSelectAllShortcut,
-} from "./clipboardShortcuts";
 import { TERMINAL_OPTIONS } from "./config";
 import { terminalRendererDebug } from "./debug";
 import { suppressQueryResponses } from "./suppressQueryResponses";
@@ -264,27 +254,6 @@ export function createTerminalInWrapper(options: CreateTerminalOptions = {}): {
 	};
 }
 
-export interface ActiveSuggestionHandle {
-	suffix: string | null;
-	onAccept: () => void;
-	onExecute: () => void;
-	onDismiss: () => void;
-	selectNext?: () => void;
-	selectPrev?: () => void;
-	hasSuggestions?: boolean;
-}
-
-export interface KeyboardHandlerOptions {
-	/** Callback for Shift+Enter (sends ESC+CR to avoid \ appearing in Claude Code while keeping line continuation behavior) */
-	onShiftEnter?: () => void;
-	/** Callback for the configured clear terminal shortcut */
-	onClear?: () => void;
-	onWrite?: (data: string) => void;
-	activeSuggestionRef?: { current: ActiveSuggestionHandle | null };
-	canOpenSuggestions?: () => boolean;
-	onOpenSuggestions?: () => void;
-}
-
 /**
  * Setup copy handler for xterm to trim trailing whitespace from copied text.
  *
@@ -326,248 +295,6 @@ export function setupCopyHandler(xterm: XTerm): () => void {
 
 	return () => {
 		element.removeEventListener("copy", handleCopy);
-	};
-}
-
-/**
- * Setup keyboard handling for xterm including:
- * - Shortcut forwarding: App hotkeys bubble to document where useAppHotkey listens
- * - Shift+Enter: Sends ESC+CR sequence (to avoid \ appearing in Claude Code while keeping line continuation behavior)
- * - Clear terminal: Uses the configured clear shortcut
- *
- * Returns a cleanup function to remove the handler.
- */
-export function setupKeyboardHandler(
-	xterm: XTerm,
-	options: KeyboardHandlerOptions = {},
-): () => void {
-	const platform =
-		typeof navigator !== "undefined" ? navigator.platform.toLowerCase() : "";
-	const isMac = platform.includes("mac");
-	const isWindows = platform.includes("win");
-
-	const handler = (event: KeyboardEvent): boolean => {
-		const noModifiers =
-			!event.metaKey && !event.ctrlKey && !event.altKey && !event.shiftKey;
-
-		if (event.key === "ArrowRight" && noModifiers && event.type === "keydown") {
-			const suggestion = options.activeSuggestionRef?.current;
-			if (suggestion?.suffix) {
-				event.preventDefault();
-				suggestion.onAccept();
-				return false;
-			}
-		}
-
-		if (event.key === "Enter" && noModifiers && event.type === "keydown") {
-			const suggestion = options.activeSuggestionRef?.current;
-			if (suggestion?.hasSuggestions) {
-				event.preventDefault();
-				suggestion.onExecute();
-				return false;
-			}
-		}
-
-		if (event.key === "ArrowDown" && noModifiers && event.type === "keydown") {
-			const suggestion = options.activeSuggestionRef?.current;
-			if (suggestion?.hasSuggestions) {
-				event.preventDefault();
-				suggestion.selectNext?.();
-				return false;
-			}
-			if (options.canOpenSuggestions?.()) {
-				event.preventDefault();
-				options.onOpenSuggestions?.();
-				return false;
-			}
-		}
-
-		if (event.key === "ArrowUp" && noModifiers && event.type === "keydown") {
-			const suggestion = options.activeSuggestionRef?.current;
-			if (suggestion?.hasSuggestions) {
-				event.preventDefault();
-				suggestion.selectPrev?.();
-				return false;
-			}
-			if (options.canOpenSuggestions?.()) {
-				event.preventDefault();
-				options.onOpenSuggestions?.();
-				return false;
-			}
-		}
-
-		if (noModifiers && (event.key === "ArrowLeft" || event.key === "Escape")) {
-			options.activeSuggestionRef?.current?.onDismiss();
-		}
-
-		const isShiftEnter =
-			event.key === "Enter" &&
-			event.shiftKey &&
-			!event.metaKey &&
-			!event.ctrlKey &&
-			!event.altKey;
-
-		if (isShiftEnter) {
-			if (event.type === "keydown" && options.onShiftEnter) {
-				event.preventDefault();
-				options.onShiftEnter();
-			}
-			return false;
-		}
-
-		const isCmdBackspace =
-			event.key === "Backspace" &&
-			event.metaKey &&
-			!event.ctrlKey &&
-			!event.altKey &&
-			!event.shiftKey;
-
-		if (isCmdBackspace) {
-			if (event.type === "keydown" && options.onWrite) {
-				event.preventDefault();
-				options.onWrite("\x15\x1b[D"); // Ctrl+U + left arrow
-			}
-			return false;
-		}
-
-		// Cmd+Left: Move cursor to beginning of line (sends Ctrl+A)
-		const isCmdLeft =
-			event.key === "ArrowLeft" &&
-			event.metaKey &&
-			!event.ctrlKey &&
-			!event.altKey &&
-			!event.shiftKey;
-
-		if (isCmdLeft) {
-			if (event.type === "keydown" && options.onWrite) {
-				event.preventDefault();
-				options.onWrite("\x01"); // Ctrl+A - beginning of line
-			}
-			return false;
-		}
-
-		// Cmd+Right: Move cursor to end of line (sends Ctrl+E)
-		const isCmdRight =
-			event.key === "ArrowRight" &&
-			event.metaKey &&
-			!event.ctrlKey &&
-			!event.altKey &&
-			!event.shiftKey;
-
-		if (isCmdRight) {
-			if (event.type === "keydown" && options.onWrite) {
-				event.preventDefault();
-				options.onWrite("\x05"); // Ctrl+E - end of line
-			}
-			return false;
-		}
-
-		// Option+Left/Right (macOS): word navigation (Meta+B / Meta+F)
-		const isOptionLeft =
-			event.key === "ArrowLeft" &&
-			event.altKey &&
-			isMac &&
-			!event.metaKey &&
-			!event.ctrlKey &&
-			!event.shiftKey;
-
-		if (isOptionLeft) {
-			if (event.type === "keydown" && options.onWrite) {
-				options.onWrite("\x1bb"); // Meta+B - backward word
-			}
-			return false;
-		}
-
-		// Option+Right: Move cursor forward by word (Meta+F)
-		const isOptionRight =
-			event.key === "ArrowRight" &&
-			event.altKey &&
-			isMac &&
-			!event.metaKey &&
-			!event.ctrlKey &&
-			!event.shiftKey;
-
-		if (isOptionRight) {
-			if (event.type === "keydown" && options.onWrite) {
-				options.onWrite("\x1bf"); // Meta+F - forward word
-			}
-			return false;
-		}
-
-		// Ctrl+Left/Right (Windows): word navigation (Meta+B / Meta+F)
-		const isCtrlLeft =
-			event.key === "ArrowLeft" &&
-			event.ctrlKey &&
-			isWindows &&
-			!event.metaKey &&
-			!event.altKey &&
-			!event.shiftKey;
-
-		if (isCtrlLeft) {
-			if (event.type === "keydown" && options.onWrite) {
-				options.onWrite("\x1bb"); // Meta+B - backward word
-			}
-			return false;
-		}
-
-		const isCtrlRight =
-			event.key === "ArrowRight" &&
-			event.ctrlKey &&
-			isWindows &&
-			!event.metaKey &&
-			!event.altKey &&
-			!event.shiftKey;
-
-		if (isCtrlRight) {
-			if (event.type === "keydown" && options.onWrite) {
-				options.onWrite("\x1bf"); // Meta+F - forward word
-			}
-			return false;
-		}
-
-		if (shouldSelectAllShortcut(event, isMac)) {
-			if (event.type === "keydown") {
-				event.preventDefault();
-				xterm.selectAll();
-			}
-			return false;
-		}
-
-		// Mirror VS Code terminal clipboard bindings so host copy/paste happens
-		// before kitty CSI-u handling in xterm consumes the command chord.
-		if (
-			shouldBubbleClipboardShortcut(event, {
-				isMac,
-				isWindows,
-				hasSelection: xterm.hasSelection(),
-			})
-		) {
-			return false;
-		}
-
-		// Terminal-reserved chords (ctrl+c/d/z/s/q) always go to xterm
-		if (isTerminalReservedEvent(event)) return true;
-
-		// CLEAR_TERMINAL is handled here (xterm needs to call onClear)
-		const clearKeys = getBinding("CLEAR_TERMINAL");
-		if (clearKeys && matchesChord(event, clearKeys)) {
-			if (event.type === "keydown" && options.onClear) {
-				options.onClear();
-			}
-			return false;
-		}
-
-		// Only bubble chords registered as app hotkeys; everything else reaches the PTY.
-		// Mirrors v2 terminal-runtime.ts:21 (VSCode terminalInstance pattern).
-		if (resolveHotkeyFromEvent(event) !== null) return false;
-
-		return true;
-	};
-
-	xterm.attachCustomKeyEventHandler(handler);
-
-	return () => {
-		xterm.attachCustomKeyEventHandler(() => true);
 	};
 }
