+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/Terminal/hooks/useTerminalSuggestion.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import { electronTrpcClient } from "renderer/lib/trpc-client";
-import type { ActiveSuggestionHandle } from "../helpers";
+import type { ActiveSuggestionHandle } from "../terminalKeyboardHandler";
 
 export interface TerminalHistorySuggestion {
 	command: string;
