+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/Terminal/terminalKeyboardHandler.ts
@@ -1,0 +1,166 @@
+import type { Terminal as XTerm } from "@xterm/xterm";
+import { resolveHotkeyFromEvent } from "renderer/hotkeys";
+import { translateLineEditChord } from "renderer/lib/terminal/line-edit-translations";
+import {
+	shouldBubbleClipboardShortcut,
+	shouldSelectAllShortcut,
+} from "./clipboardShortcuts";
+
+// FORK NOTE: ActiveSuggestionHandle and suggestion-related options were
+// originally defined in helpers.ts and are now co-located here alongside
+// setupKeyboardHandler (upstream #3724 migration).
+export interface ActiveSuggestionHandle {
+	suffix: string | null;
+	onAccept: () => void;
+	onExecute: () => void;
+	onDismiss: () => void;
+	selectNext?: () => void;
+	selectPrev?: () => void;
+	hasSuggestions?: boolean;
+}
+
+export interface KeyboardHandlerOptions {
+	/** Callback for Shift+Enter (sends ESC+CR to avoid \ appearing in Claude Code while keeping line continuation behavior) */
+	onShiftEnter?: () => void;
+	onWrite?: (data: string) => void;
+	// FORK NOTE: shell history suggestion integration (PR #<suggestion-feature>).
+	// Not present in upstream — these drive the TerminalSuggestion overlay.
+	// Follow-up: apply to v2 terminal if needed.
+	activeSuggestionRef?: { current: ActiveSuggestionHandle | null };
+	canOpenSuggestions?: () => boolean;
+	onOpenSuggestions?: () => void;
+}
+
+/**
+ * Setup keyboard handling for xterm including:
+ * - Shortcut forwarding: App hotkeys bubble to document where useAppHotkey listens
+ * - Shift+Enter: Sends ESC+CR sequence (to avoid \ appearing in Claude Code while keeping line continuation behavior)
+ * - Line-edit chord translations (shared with v2 via translateLineEditChord)
+ * - Shell history suggestion navigation (fork-only)
+ *
+ * Returns a cleanup function to remove the handler.
+ */
+export function setupKeyboardHandler(
+	xterm: XTerm,
+	options: KeyboardHandlerOptions = {},
+): () => void {
+	const platform =
+		typeof navigator !== "undefined" ? navigator.platform.toLowerCase() : "";
+	const isMac = platform.includes("mac");
+	const isWindows = platform.includes("win");
+
+	const handler = (event: KeyboardEvent): boolean => {
+		// Match v2: registered app hotkeys must escape xterm before terminal
+		// translations or macOS Cmd bubbling can consume them.
+		if (resolveHotkeyFromEvent(event) !== null) return false;
+
+		const noModifiers =
+			!event.metaKey && !event.ctrlKey && !event.altKey && !event.shiftKey;
+
+		// FORK NOTE: suggestion overlay key handling — intercepts arrow keys
+		// and Enter when a suggestion is active before the shell sees them.
+		if (event.key === "ArrowRight" && noModifiers && event.type === "keydown") {
+			const suggestion = options.activeSuggestionRef?.current;
+			if (suggestion?.suffix) {
+				event.preventDefault();
+				suggestion.onAccept();
+				return false;
+			}
+		}
+
+		if (event.key === "Enter" && noModifiers && event.type === "keydown") {
+			const suggestion = options.activeSuggestionRef?.current;
+			if (suggestion?.hasSuggestions) {
+				event.preventDefault();
+				suggestion.onExecute();
+				return false;
+			}
+		}
+
+		if (event.key === "ArrowDown" && noModifiers && event.type === "keydown") {
+			const suggestion = options.activeSuggestionRef?.current;
+			if (suggestion?.hasSuggestions) {
+				event.preventDefault();
+				suggestion.selectNext?.();
+				return false;
+			}
+			if (options.canOpenSuggestions?.()) {
+				event.preventDefault();
+				options.onOpenSuggestions?.();
+				return false;
+			}
+		}
+
+		if (event.key === "ArrowUp" && noModifiers && event.type === "keydown") {
+			const suggestion = options.activeSuggestionRef?.current;
+			if (suggestion?.hasSuggestions) {
+				event.preventDefault();
+				suggestion.selectPrev?.();
+				return false;
+			}
+			if (options.canOpenSuggestions?.()) {
+				event.preventDefault();
+				options.onOpenSuggestions?.();
+				return false;
+			}
+		}
+
+		if (noModifiers && (event.key === "ArrowLeft" || event.key === "Escape")) {
+			options.activeSuggestionRef?.current?.onDismiss();
+		}
+
+		const isShiftEnter =
+			event.key === "Enter" &&
+			event.shiftKey &&
+			!event.metaKey &&
+			!event.ctrlKey &&
+			!event.altKey;
+
+		if (isShiftEnter) {
+			if (event.type === "keydown" && options.onShiftEnter) {
+				event.preventDefault();
+				options.onShiftEnter();
+			}
+			return false;
+		}
+
+		const translation = translateLineEditChord(event, { isMac, isWindows });
+		if (translation !== null) {
+			if (event.type === "keydown" && options.onWrite) {
+				event.preventDefault();
+				options.onWrite(translation);
+			}
+			return false;
+		}
+
+		if (shouldSelectAllShortcut(event, isMac)) {
+			if (event.type === "keydown") {
+				event.preventDefault();
+				xterm.selectAll();
+			}
+			return false;
+		}
+
+		// Mirror VS Code terminal clipboard bindings so host copy/paste happens
+		// before kitty CSI-u handling in xterm consumes the command chord.
+		if (
+			shouldBubbleClipboardShortcut(event, {
+				isMac,
+				isWindows,
+				hasSelection: xterm.hasSelection(),
+			})
+		) {
+			return false;
+		}
+
+		// Default: let xterm process unhandled keys, including terminal-reserved
+		// chords like ctrl+c/d/z/s/q.
+		return true;
+	};
+
+	xterm.attachCustomKeyEventHandler(handler);
+
+	return () => {
+		xterm.attachCustomKeyEventHandler(() => true);
+	};
+}
